casts to fix warnings in MSVC x64, thanks Willis
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@6045 d7cf8633-e32d-0410-b094-e92efae38249

+++ host_cmd.c
@@ -2230,7 +2230,7 @@ void Host_Packet_f (void) // credit: taken from QuakeWorld
 	out = send+4;
 	send[0] = send[1] = send[2] = send[3] = 0xff;
 
-	l = strlen (in);
+	l = (int)strlen (in);
 	for (i=0 ; i<l ; i++)
 	{
 		if (out >= send + sizeof(send) - 1)
